Fix int to long mismatch

+++ Server/src/common/infrastructure/db/SqliteDatabase.cpp
@@ -236,7 +236,7 @@ long SqliteDatabase::addNewUser(const std::string &username, const std::string &
                                         const std::string &email, const std::string &phone,
                                         const std::string &birthdate, const std::optional<std::string> &address) const
 {
-    const int id = *queryIds(
+    const long id = *queryIds(
         this->_preppedStatements.at("addNewUser"),
         {
             username,
